Use the new method createEntryPoint to install correctly gobsapi.php

+++ gobsapi/install/install.php
@@ -11,11 +11,18 @@ class gobsapiModuleInstaller extends jInstallerModule
 {
     public function install()
     {
+        if (method_exists($this, 'createEntryPoint')) {
+            $this->createEntryPoint('gobsapi.php', 'gobsapi/config.ini.php', 'gobsapi', 'classic');
+        }
+        else {
+            // Copy configuration file
+            $gobs_config_target = jApp::configPath('gobsapi.ini.php');
+            if (!file_exists($gobs_config_target)) {
+                $this->copyFile('config/gobsapi.ini.php.dist', $gobs_config_target);
+            }
 
-        // Copy configuration file
-        $gobs_config_target = jApp::configPath('gobsapi.ini.php');
-        if (!file_exists($gobs_config_target)) {
-            $this->copyFile('config/gobsapi.ini.php.dist', $gobs_config_target);
+            //deprecated and not safe
+            $this->copyFile('gobsapi.php', jApp::wwwPath('gobsapi.php'));
         }
     }
 }
